+++ project_composer/manifest.py
@@ -12,14 +12,55 @@ import tomli
 from .exceptions import ComposerManifestError
 
 
-class TextContentConfig:
+class BaseConfig:
     """
-    Dedicated manifest part model to TextContent operations.
+    Configuration class abstract.
+
+    Attributes:
+        _FIELDS (list): List of payload attribute names. Only those attributes will
+            be available to dump.
+    """
+    _FIELDS = []
+
+    def get_fields(self):
+        """
+        Return model payload attribute names.
+
+        Returns:
+            list: List name string for each payload attribute.
+        """
+        return self._FIELDS
+
+    def to_dict(self):
+        """
+        Dump manifest values as Python dictionnary.
+
+        Recursively walk in ``to_dict`` method of item that are a children of
+        BaseConfig.
+
+        Returns:
+            dict: Dictionnary of all field values, including the "requirements"
+            ones.
+        """
+        content = {}
+
+        for name in self.get_fields():
+            attr = getattr(self, name)
+            if isinstance(attr, BaseConfig):
+                content[name] = attr.to_dict()
+            else:
+                content[name] = getattr(self, name)
+
+        return content
+
+
+class TextContentConfig(BaseConfig):
+    """
+    Configuration class for TextContent alike plugin.
     """
     _FIELDS = [
         "application_label",
         "application_divider",
-        "dump",
         "introduction",
         "source_filename",
         "template",
@@ -30,10 +71,9 @@ class TextContentConfig:
     )
     _DEFAULT_CONTENT_FILENAME = "source.txt"
 
-    def __init__(self, dump=None, minimal=False, application_label=None,
-                 application_divider=None, template=None, introduction=None,
-                 source_filename=None):
-        self.dump = dump
+    def __init__(self, application_label=None, application_divider=None, template=None,
+                 introduction=None, source_filename=None):
+        # TODO: Attributes should be set from _FIELDS
         self.application_label = application_label
         self.application_divider = application_divider
         self.template = template
@@ -49,15 +89,17 @@ class TextContentConfig:
 
 class RequirementsConfig(TextContentConfig):
     """
-    Requirements class so options are reachable as proper object attributes in their
-    own namespace
+    Requirements file plugin.
     """
     _DEFAULT_CONTENT_FILENAME = "requirements.txt"
 
 
-class Manifest:
+class Manifest(BaseConfig):
     """
     Manifest model.
+
+    NOTE: Currently adding new fields will need some declarations changes here and
+    there because there is not strong declaration yet as in real data model.
 
     Arguments:
         name (string): The manifest title name.
@@ -75,8 +117,19 @@ class Manifest:
     Attributes:
         name (string): The manifest title name.
         apps (list): A list of application module names for enabled application.
-        requirements (RequirementsConfig): Requirements specific options.
+        requirements (dict or RequirementsConfig): Requirements specific options.
+            Either as RequirementsConfig object or a dict of values respecting the
+            RequirementsConfig attributes.
     """
+    # Payload fields declaration
+    _FIELDS = [
+        "name",
+        "apps",
+        "repository",
+        "syspaths",
+        "requirements",
+    ]
+
     def __init__(self, name, apps, repository=None, syspaths=None, requirements=None):
         self.name = name
         self.apps = apps
@@ -86,7 +139,10 @@ class Manifest:
         # For possible future new parts, we would need about a plugins namespace where
         # to register 'requirements' and other parts as a plugin.
         _reqs = requirements or {}
-        self.requirements = RequirementsConfig(**_reqs)
+        if isinstance(_reqs, RequirementsConfig):
+            self.requirements = _reqs
+        else:
+            self.requirements = RequirementsConfig(**_reqs)
 
         super().__init__()
 
@@ -136,11 +192,15 @@ class Manifest:
             # Enforce Path object
             source = Path(source)
 
+            # JSON format is the simpliest structure which fit exactly to the manifest
+            # structure
             if source.name.endswith(".json"):
                 content = json.loads(source.read_text())
+            # TOML format is more complex than manifest structure
             elif source.name.endswith(".toml"):
                 loaded = tomli.loads(source.read_text())
 
+                # Check required structure
                 if (
                     "tool" not in loaded or
                     "project_composer" not in loaded["tool"]
@@ -150,13 +210,19 @@ class Manifest:
                         "fill base options."
                     )
 
+                composer_section = loaded["tool"]["project_composer"]
+
+                # Build payload
                 content = {
-                    "name": loaded.get("project", {}).get("name", ""),
-                    "apps": loaded["tool"]["project_composer"].get("apps", None),
-                    "requirements": loaded["tool"]["project_composer"].get(
-                        "requirements"
-                    ),
+                    name: composer_section.get(name, None)
+                    for name in cls._FIELDS
                 }
+
+                # If there is no composer config name, try to use the TOML project one
+                if not content["name"]:
+                    content["name"] = loaded.get("project", {}).get("name", "")
+
+            # No recognized format
             else:
                 raise ComposerManifestError(
                     "Unable to guess the manifest file format. Please suffix your "