+++ tests/0002_manifest.py
@@ -6,6 +6,45 @@ from project_composer.exceptions import ComposerManifestError
 from project_composer.manifest import RequirementsConfig, Manifest
 
 
+def test_manifest_to_dict():
+    """
+    Manifest object should correctly dump all its field values to a dictionnary.
+    """
+    manifest = Manifest(
+        "foo",
+        ["bar"],
+        repository="plop.plip",
+        syspaths=["flip", "flop", "flip"],
+        requirements=RequirementsConfig(
+            application_label="label",
+            application_divider="div",
+            template="template.txt",
+            source_filename="source.txt",
+            introduction="intro",
+        ),
+    )
+
+    content = manifest.to_dict()
+
+    assert content == {
+        "name": "foo",
+        "apps": ["bar"],
+        "repository": "plop.plip",
+        "syspaths": [
+            "flip",
+            "flop",
+            "flip"
+        ],
+        "requirements": {
+            "application_label": "label",
+            "application_divider": "div",
+            "introduction": "intro",
+            "source_filename": "source.txt",
+            "template": "template.txt"
+        }
+    }
+
+
 def test_manifest_load_invalid_format(pytester, settings):
     """
     Manifest filepath must have the right extension to guess its format
@@ -54,49 +93,54 @@ def test_manifest_load_invalid_json_format(settings, manifest_filename, expected
     assert exc_info.value.args[0].startswith(expected) is True
 
 
-@pytest.mark.parametrize(
-    "manifest_filename, expected_name, expected_apps, expected_requirement",
-    [
-        (
-            "basic.json",
-            "Sample",
-            [
+@pytest.mark.parametrize("manifest_filename, expected", [
+    (
+        "basic.json",
+        {
+            "name": "Sample",
+            "apps": [
                 "foo",
                 "bar",
                 "ping",
                 "nope",
             ],
-            {
+            "repository": "apps_structure",
+            "syspaths": [],
+            "requirements": {
                 "application_label": None,
                 "application_divider": None,
-                "dump": None,
                 "introduction": RequirementsConfig._DEFAULT_INTRO,
                 "source_filename": RequirementsConfig._DEFAULT_CONTENT_FILENAME,
                 "template": None,
-            },
-        ),
-        (
-            "basic.toml",
-            "Sample",
-            [
+            }
+        },
+    ),
+    (
+        "basic.toml",
+        {
+            "name": "Sample",
+            "apps": [
                 "foo",
                 "bar",
                 "ping",
                 "nope",
             ],
-            {
+            "repository": "apps_structure",
+            "syspaths": [],
+            "requirements": {
                 "application_label": None,
                 "application_divider": None,
-                "dump": None,
                 "introduction": RequirementsConfig._DEFAULT_INTRO,
                 "source_filename": RequirementsConfig._DEFAULT_CONTENT_FILENAME,
                 "template": None,
-            },
-        ),
-        (
-            "full.json",
-            "Full manifest options",
-            [
+            }
+        },
+    ),
+    (
+        "full.json",
+        {
+            "name": "Full manifest JSON",
+            "apps": [
                 "bar",
                 "dummy",
                 "empty",
@@ -105,19 +149,22 @@ def test_manifest_load_invalid_json_format(settings, manifest_filename, expected
                 "pong",
                 "nope",
             ],
-            {
+            "repository": "apps_structure",
+            "syspaths": [],
+            "requirements": {
                 "application_label": "# {name}\n",
                 "application_divider": "\n",
-                "dump": "output.txt",
                 "introduction": None,
                 "source_filename": RequirementsConfig._DEFAULT_CONTENT_FILENAME,
                 "template": "requirements_template.txt",
-            },
-        ),
-        (
-            "full.toml",
-            "Full manifest options",
-            [
+            }
+        },
+    ),
+    (
+        "full.toml",
+        {
+            "name": "Full manifest TOML",
+            "apps": [
                 "bar",
                 "dummy",
                 "empty",
@@ -126,19 +173,19 @@ def test_manifest_load_invalid_json_format(settings, manifest_filename, expected
                 "pong",
                 "nope",
             ],
-            {
+            "repository": "apps_structure",
+            "syspaths": [],
+            "requirements": {
                 "application_label": "# {name}\n",
                 "application_divider": "\n",
-                "dump": "output.txt",
                 "introduction": None,
                 "source_filename": RequirementsConfig._DEFAULT_CONTENT_FILENAME,
                 "template": "requirements_template.txt",
-            },
-        ),
-    ]
-)
-def test_manifest_load(settings, manifest_filename, expected_name, expected_apps,
-                       expected_requirement):
+            }
+        },
+    ),
+])
+def test_manifest_load(settings, manifest_filename, expected):
     """
     Manifest should load a valid file and correctly fill its attribute from content,
     no matter what format backend is used.
@@ -146,13 +193,4 @@ def test_manifest_load(settings, manifest_filename, expected_name, expected_apps
     manifest_source = settings.fixtures_path / "manifests" / manifest_filename
     manifest = Manifest.load(manifest_source)
 
-    assert manifest.name == expected_name
-    assert manifest.apps == expected_apps
-
-    # It's easier to test requirements config as dict against expected data
-    requirements_kwargs = {
-        k: getattr(manifest.requirements, k)
-        for k in RequirementsConfig._FIELDS
-    }
-
-    assert requirements_kwargs == expected_requirement
+    assert manifest.to_dict() == expected